+++ src/main/java/de/yuga/spacebattle/backend/repositories/constructables/buildings/ConstructionRepository.java
@@ -5,8 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
+import java.util.Set;
 
 public interface ConstructionRepository extends JpaRepository<Construction, Integer>, CustomConstructionRepository {
 
@@ -21,4 +23,40 @@ public interface ConstructionRepository extends JpaRepository<Construction, Inte
     @Nullable
     @Query("SELECT c FROM Construction c WHERE c.planet.id = :idPlanet AND c.level > c.operationalLevel")
     List<Construction> findInoperationalForPlanet(@Param("idPlanet") final int idPlanet);
+
+    @Nullable
+    @Query("SELECT c FROM Construction c WHERE c.planet.id = :idPlanet")
+    Set<Construction> findAllConstructionsOnPlanet(final int idPlanet);
+
+    @Nullable
+    @Query("SELECT c FROM Construction c LEFT JOIN FETCH c.jobs j WHERE c.planet.id = :idPlanet")
+    Set<Construction> findAllConstructionsOnPlanetWithJobs(final int idPlanet);
+
+    @Query("SELECT CASE WHEN (COUNT(c) > 0) THEN TRUE ELSE FALSE END FROM Construction c " +
+            "WHERE c.planet.id = :idPlanet " +
+            "AND c.building.productionType.productionTarget = :productionTarget " +
+            "AND c.building.productionType.productionCategory = :productionCategory")
+    boolean hasPlanetProductionType(final int idPlanet, @Nonnull final String productionTarget, @Nonnull final String productionCategory);
+
+    @Query("SELECT CASE WHEN (COUNT(c) > 0) THEN TRUE ELSE FALSE END FROM Construction c LEFT JOIN c.jobs j " +
+            "WHERE c.planet.id = :idPlanet " +
+            "AND c.building.productionType.productionTarget = :productionTarget " +
+            "AND c.building.productionType.productionCategory = :productionCategory " +
+            "AND j.isDeleted = false ")
+    boolean isActiveJobPresentForTargetAtPlanet(final int idPlanet, @Nonnull final String productionTarget, @Nonnull final String productionCategory);
+
+    @Nullable
+    @Query("SELECT c FROM Construction c " +
+            "LEFT JOIN FETCH c.jobs j " +
+            "WHERE c.planet.id = :idPlanet " +
+            "AND c.building.id = :idBuilding")
+    Construction findByPlanetAndBuilding(int idPlanet, int idBuilding);
+
+    @Nullable
+    @Query("SELECT c FROM Construction c " +
+            "LEFT JOIN FETCH c.jobs j " +
+            "WHERE c.planet.id = :idPlanet " +
+            "AND c.building.productionType.productionTarget = :productionTarget " +
+            "AND c.building.productionType.productionCategory = :productionCategory")
+    Construction findByPlanetProductionType(final int idPlanet, @Nonnull final String productionTarget, @Nonnull final String productionCategory);
 }