+++ src/main/java/de/yuga/spacebattle/backend/services/orbitals/PlanetService.java
@@ -106,6 +106,7 @@ public class PlanetService {
 
     @Nonnull
     List<Planet> findAllColonizedByWithResearchLab(final int idUser) {
+        /* fixme pretty bad performance */
         return Objects.requireNonNullElse(planetRepository.findAllColonizedByWithResearchLab(idUser), List.of());
     }